Added some comments
The code is fairly easy to follow but I added some comments. The comments are not perfect but should allow for better understanding especially where I used "zero" value of a variable to reference its state while also holding place values. (and subtracting one after I knew that it was a non zero to have zero-"x" values.